+++ packages/at_client/lib/at_client.dart
@@ -32,6 +32,8 @@ export 'package:at_auth/at_auth.dart'
         NamespacePermission,
         WrittenAtKeysIo;
 export 'package:at_auth/at_auth_io.dart' show FileAtKeysIo;
+export 'package:at_lookup/at_lookup_io.dart'
+    show AtLookUpFactory, secureSocketLookUps, AtCommandExecutor;
 export 'package:at_chops/at_chops.dart' show SigningAlgoType;
 export 'package:at_client/src/response/at_notification.dart';
 export 'package:at_client/src/response/enrollment.dart';