+++ packages/at_onboarding_cli/lib/src/onboard/at_onboarding_service_impl.dart
@@ -70,7 +70,8 @@ class AtOnboardingServiceImpl implements AtOnboardingService {
           preference: atOnboardingPreference,
           storage: atOnboardingPreference.storageFor(_atSign),
           serviceFactory: atServiceFactory,
-          atLookUp: _atLookUp);
+          atLookUp: _atLookUp,
+          lookUps: atOnboardingPreference.lookUps);
     } on AtOpenRefusedException catch (e) {
       logger.warning(e.message);
       return false;